Fixed tools.classpath.
We don't ship a narya-tools artifact to Maven Central, so we need to specify a
dependency on narya proper (which includes samskivert and guava) and add a
manual dependency on javassist (which Narya does not export as a dependency).

I wish Maven had build-time-only dependencies (or 'tools' dependencies), then I
could just export javassist as such for Narya and be done with it.